Transcendence/requirements.txt
Django==4.2.9
daphne
djangorestframework==3.14.0
pyotp==2.9.0
qrcode==7.4.2
pyJWT==2.8.0
requests==2.31.0
channels==4.0.0
cookies==2.2.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
asgiref==3.7.2
sqlparse==0.4.4
typing_extensions==4.10.0
gunicorn==20.1.0
tzdata
psycopg2-binary